// WADNR.Web/src/app/shared/generated/model/project-type-detail.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectTypeDetail { 
    ProjectTypeID?: number;
    TaxonomyBranchID?: number;
    ProjectTypeName?: string | null;
    ProjectTypeDescription?: string | null;
    ProjectTypeCode?: string | null;
    ThemeColor?: string | null;
    ProjectTypeSortOrder?: number | null;
    LimitVisibilityToAdmin?: boolean;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectTypeDetailForm { 
    ProjectTypeID?: FormControl<number>;
    TaxonomyBranchID?: FormControl<number>;
    ProjectTypeName?: FormControl<string>;
    ProjectTypeDescription?: FormControl<string>;
    ProjectTypeCode?: FormControl<string>;
    ThemeColor?: FormControl<string>;
    ProjectTypeSortOrder?: FormControl<number>;
    LimitVisibilityToAdmin?: FormControl<boolean>;
}

export class ProjectTypeDetailFormControls { 
    public static ProjectTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static TaxonomyBranchID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProjectTypeName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProjectTypeDescription = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProjectTypeCode = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ThemeColor = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProjectTypeSortOrder = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static LimitVisibilityToAdmin = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}